+++ packages/dialtone-tokens/release-ci.config.cjs
@@ -1,8 +1,11 @@
+/* eslint-disable no-template-curly-in-string */
 const name = 'dialtone-tokens';
 const srcRoot = `packages/${name}`;
 
+/**
+ * @type {import('semantic-release').GlobalConfig}
+ */
 module.exports = {
-  extends: 'release.config.base.js',
   pkgRoot: srcRoot,
   tagFormat: name + '/v${version}',
   commitPaths: [`${srcRoot}/*`],
@@ -10,16 +13,36 @@ module.exports = {
     ['@semantic-release/commit-analyzer', {
       preset: 'angular',
       releaseRules: [
-        {type: 'refactor', release: 'patch'},
+        { type: 'refactor', release: 'patch' },
       ],
     }],
     ['@semantic-release/release-notes-generator', {
       config: '@dialpad/conventional-changelog-angular',
     }],
+    ['@dialpad/semantic-release-changelog-json', {
+      changelogFile: `${srcRoot}/CHANGELOG.md`,
+      changelogJsonFile: `${srcRoot}/CHANGELOG.json`,
+    }],
+    ['@semantic-release/changelog', { changelogFile: `${srcRoot}/CHANGELOG.md` }],
+    ['@semantic-release/npm', { npmPublish: false }],
+    ['@semantic-release/git', {
+      assets: [
+        `${srcRoot}/CHANGELOG.md`,
+        `${srcRoot}/CHANGELOG.json`,
+        `${srcRoot}/package.json`,
+      ],
+      message: `chore(release): NO-JIRA ${name}` +
+        '/v${nextRelease.version}\n\n${nextRelease.notes}',
+    }],
+    ['@semantic-release/exec', {
+      prepareCmd: './gradlew setProperties -Pversion=${nextRelease.version} && echo \'${nextRelease.version}\' > ./dist_ios/VERSION && git add -A && git commit -m \'chore(release): NO-JIRA ' + name + '/v${nextRelease.version} gradle\' && git push',
+      execCwd: srcRoot,
+    }],
     '@semantic-release/github',
   ],
   branches: [
-    'production',
+    'staging',
+    'next',
     {
       name: 'beta',
       prerelease: true,