+++ packages/dialtone-vue2/recipes/conversation_view/message_input/message_input_default.story.vue
@@ -3,7 +3,7 @@
     <div class="d-h264">
       <dt-recipe-message-input
         ref="input"
-        v-model="value"
+        v-model="$attrs.value"
         :input-aria-label="$attrs.inputAriaLabel"
         :auto-focus="$attrs.autoFocus"
         :allow-blockquote="$attrs.allowBlockquote"
@@ -21,6 +21,7 @@
         :max-height="$attrs.maxHeight"
         :mention-suggestion="$attrs.mentionSuggestion"
         :channel-suggestion="$attrs.channelSuggestion"
+        :slash-command-suggestion="$attrs.slashCommandSuggestion"
         :show-emoji-picker="$attrs.showEmojiPicker"
         :emoji-picker-props="$attrs.emojiPickerProps"
         :emoji-tooltip-message="$attrs.emojiTooltipMessage"
@@ -35,6 +36,7 @@
         @input="$attrs.onInput"
         @select-media="$attrs.onSelectMedia"
         @selected-emoji="$attrs.onSelectedEmoji"
+        @selected-command="$attrs.onSelectedCommand"
         @skin-tone="$attrs.onSkinTone"
         @add-media="$attrs.onAddMedia"
         @paste-media="$attrs.onPasteMedia"
@@ -51,10 +53,5 @@ import DtRecipeMessageInput from './message_input.vue';
 export default {
   name: 'DtRecipeMessageInputDefault',
   components: { DtRecipeMessageInput },
-  data () {
-    return {
-      value: this.$attrs.value,
-    };
-  },
 };
 </script>