+++ packages/dialtone-vue2/recipes/leftbar/general_row/general_row_variants.story.vue
@@ -36,6 +36,40 @@
     </div>
     <div>
       <h3>
+        Channel with unread messages
+      </h3>
+      <dt-recipe-general-row
+        type="channels"
+        unread-count="1"
+        :has-unreads="true"
+        description="Channel name"
+      />
+    </div>
+    <div>
+      <h3>
+        Channel only show unread mentioned count
+      </h3>
+      <dt-recipe-general-row
+        type="channels"
+        unread-mention-count="1"
+        :has-unreads="true"
+        description="Channel name"
+      />
+    </div>
+    <div>
+      <h3>
+        Channel show both unread and unread mention count
+      </h3>
+      <dt-recipe-general-row
+        type="channels"
+        unread-count="4"
+        :has-unreads="true"
+        unread-mention-count="1"
+        description="Channel name"
+      />
+    </div>
+    <div>
+      <h3>
         Contact Center with DND and unread count
       </h3>
       <dt-recipe-general-row