+++ release-ci.config.cjs
@@ -2,8 +2,10 @@
 const name = 'dialtone';
 const srcRoot = '.';
 
+/**
+ * @type {import('semantic-release').GlobalConfig}
+ */
 module.exports = {
-  extends: 'release.config.base.js',
   pkgRoot: srcRoot,
   tagFormat: name + '/v${version}',
   commitPaths: [`${srcRoot}/*`],
@@ -17,10 +19,26 @@ module.exports = {
     ['@semantic-release/release-notes-generator', {
       config: '@dialpad/conventional-changelog-angular',
     }],
+    ['@dialpad/semantic-release-changelog-json', {
+      changelogFile: `${srcRoot}/CHANGELOG.md`,
+      changelogJsonFile: `${srcRoot}/CHANGELOG.json`,
+    }],
+    ['@semantic-release/changelog', { changelogFile: `${srcRoot}/CHANGELOG.md` }],
+    ['@semantic-release/npm', { npmPublish: false }],
+    ['@semantic-release/git', {
+      assets: [
+        `${srcRoot}/CHANGELOG.md`,
+        `${srcRoot}/CHANGELOG.json`,
+        `${srcRoot}/package.json`,
+        `${srcRoot}/pnpm-lock.yaml`,
+      ],
+      message: `chore(release): NO-JIRA ${name}` +
+        '/v${nextRelease.version}\n\n${nextRelease.notes}',
+    }],
     '@semantic-release/github',
   ],
   branches: [
-    'production',
+    'staging',
     'next',
     {
       name: 'beta',